1.8.0
recruits 1.8.0:

+ fixed friendlyfire bug
+ fixed dialogues bug
+ fixed clear target dont work
+ fixed recruits that are holding pos. dont flee when a tnt nearby is primed
+ fixed pillagers could have a shield while having a crossbow

+ added mount command
	- recruits will mount the Vehicle/Mob you are looking at
	- compatible with smallships and camels

+ added dismount command
+ added whitelist for entities to mount (configurable)
+ added escort command
	- recruits can escort a Mob to secure its safty
	- recruits will attack incoming danger for the escorting Mob

+ added Block place/break will aggro ercruits (configurable)
	- if the recruits are in aggressive mode
	and a enemy player place/break a block the recruits will attack this player immediately

+ added banner/scull layer to recruits
+ added the targeted group to commands in chat
+ added quick equip to recruit inventory menu

*improved code

https://www.curseforge.com/minecraft/mc-mods/recruits/files

+++ src/main/java/com/talhanation/recruits/entities/ai/RecruitFollowOwnerGoal.java
@@ -42,7 +42,7 @@ public class RecruitFollowOwnerGoal extends Goal {
             return false;
         } else if (!this.recruitEntity.getShouldFollow()) {
             return false;
-        } else if (this.recruitEntity.distanceToSqr(owner) < (double)(this.startDistance * this.startDistance)) {
+        } else if (this.recruitEntity.distanceToSqr(owner) < (this.startDistance * this.startDistance)) {
             return false;
         }
         else {
